Added criteria local data verification(2.7)
This change proposes that applications ensure that the local data generated on one device, cannot be duplicated and utilised by the same application on another device. Applications are required to sign local data in such a way that local data is linked to a device to prevent the duplication and abuse of local data on rogue devices.