Accept certificates from trusted CA's
It is not really clear what a "valid" CA is. We want to allow private PKI where someone sets up his own CA, but we don't want to accept certificates signed with the attacker's CA. It can be argued that both are "valid" CA's. Change this to "trusted". The application can use the built-in CA store and trust the same CA's that the OS trusts, or choose to trust only some CA's, or trust only its own private CA.

See also #105.